mudanças codigo coesão
